Rewrite the integration tests
- Use Auth0 for authentication -> allow running against staging or production
- Have configurations for different environments
- Use a separate 'CharlesClient' for issuing the API requests

Some of the tests have not yet been ported so they are still preserved.
